% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catboost.R
\name{axe-catboost.Model}
\alias{axe-catboost.Model}
\alias{axe_call.catboost.Model}
\alias{axe_ctrl.catboost.Model}
\alias{axe_data.catboost.Model}
\alias{axe_env.catboost.Model}
\alias{axe_fitted.catboost.Model}
\title{Axing a catboost.Model.}
\usage{
\method{axe_call}{catboost.Model}(x, verbose = FALSE, ...)

\method{axe_ctrl}{catboost.Model}(x, verbose = FALSE, ...)

\method{axe_data}{catboost.Model}(x, verbose = FALSE, ...)

\method{axe_env}{catboost.Model}(x, verbose = FALSE, ...)

\method{axe_fitted}{catboost.Model}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed catboost.Model object.
}
\description{
catboost.Model objects are created from the \pkg{catboost} package.
These models store minimal training artifacts but do include
pre-computed feature importances that can be removed.
}
\examples{
\dontshow{if (rlang::is_installed(c("catboost", "parsnip", "bonsai"))) withAutoprint(\{ # examplesIf}
library(parsnip)
library(bonsai)

mod <- boost_tree(trees = 10) \%>\%
  set_engine("catboost", verbose = 0) \%>\%
  set_mode("classification") \%>\%
  fit(Species ~ ., data = iris)

# Extract the underlying catboost model
catboost_model <- mod$fit

butchered <- butcher(catboost_model, verbose = TRUE)

# Predictions still work
mod$fit <- butchered
predict(mod, iris)
\dontshow{\}) # examplesIf}
}
