% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classbagg.R
\name{axe-classbagg}
\alias{axe-classbagg}
\alias{axe_call.classbagg}
\alias{axe_data.classbagg}
\alias{axe_env.classbagg}
\title{Axing a classbagg object.}
\usage{
axe_call.classbagg(x, ...)

axe_data.classbagg(x, ...)

axe_env.classbagg(x, ...)
}
\arguments{
\item{x}{classbagg object}

\item{...}{any additional arguments related to axing}
}
\value{
axed classbagg object
}
\description{
Axing a classbagg object.

Remove the call. Calls located within each subtree is also removed.

Remove training data. There are also responses stored as either a factor
vector of class labels, a vector of numerical values, or an object of class
\code{Surv} that can also be removed from \code{x$y} for consistency.

Remove environments. Model objects of this type include references to
environments in each step of the recipe, and thus must also be
removed. Note that environments that result from \code{srcref} are
not axed.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(ipred)))
suppressWarnings(suppressMessages(library(rpart)))
suppressWarnings(suppressMessages(library(MASS)))

# Load data
data("GlaucomaM", package = "TH.data")

classbagg_fit <- bagging(Class ~ ., data = GlaucomaM, coob = TRUE)

butcher(classbagg_fit)
}
