% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{axe-cv.glmnet}
\alias{axe-cv.glmnet}
\alias{axe_fitted.cv.glmnet}
\title{Axing an cv.glmnet.}
\usage{
\method{axe_fitted}{cv.glmnet}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed cv.glmnet object.
}
\description{
cv.glmnet objects are created from carrying out k-fold
cross-validation from the \pkg{glmnet} package.
}
\examples{
# Load libraries
library(glmnet)
n <- 500;p <- 30
nzc <- trunc(p/10)
x <- matrix(rnorm(n*p), n, p)
beta3 <- matrix(rnorm(30), 10, nzc)
beta3 <- rbind(beta3, matrix(0, p-10, nzc))
f3 <- x \%*\% beta3
p3 <- exp(f3)
p3 <- p3/apply(p3, 1, sum)
g3 <- rmult(p3)
set.seed(10101)
cvfit <- cv.glmnet(x, g3, family="multinomial", keep = TRUE)

butcher(cvfit)
}
