% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{axe-default}
\alias{axe-default}
\alias{axe.default}
\alias{axe_call.default}
\alias{axe_ctrl.default}
\alias{axe_data.default}
\alias{axe_env.default}
\alias{axe_fitted.default}
\alias{axe_misc.default}
\alias{predict.butcher_default}
\title{Axing for default inputs.}
\usage{
\method{axe}{default}(x, ...)

\method{axe_call}{default}(x, ...)

\method{axe_ctrl}{default}(x, ...)

\method{axe_data}{default}(x, ...)

\method{axe_env}{default}(x, ...)

\method{axe_fitted}{default}(x, ...)

\method{axe_misc}{default}(x, ...)

\method{predict}{butcher_default}(x, ...)
}
\description{
These default axe functions treat the model object as if
nothing needs to be axed.
}
