% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{axe-earth}
\alias{axe-earth}
\alias{axe_call.earth}
\alias{axe_data.earth}
\alias{axe_fitted.earth}
\title{Axing an earth object.}
\usage{
\method{axe_call}{earth}(x, ...)

\method{axe_data}{earth}(x, ...)

\method{axe_fitted}{earth}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
earth objects are created from the \code{earth} package, which
is leveraged to do multivariate adapative regression splines.
This is where all the earth specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))

# Create model and fit
earth_fit <- mars(mode = "regression") \%>\%
  set_engine("earth") \%>\%
  fit(Volume ~ ., data = trees)

butcher(earth_fit)

}
