% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexsurvreg.R
\name{axe-flexsurvreg}
\alias{axe-flexsurvreg}
\alias{axe.flexsurvreg}
\alias{axe_call.flexsurvreg}
\alias{axe_ctrl.flexsurvreg}
\alias{axe_data.flexsurvreg}
\alias{axe_env.flexsurvreg}
\alias{axe_fitted.flexsurvreg}
\alias{axe_misc.flexsurvreg}
\alias{predict.butcher_flexsurvreg}
\title{Axing an flexsurvreg.}
\usage{
axe.flexsurvreg(x, ...)

\method{axe_call}{flexsurvreg}(x, ...)

\method{axe_ctrl}{flexsurvreg}(x, ...)

\method{axe_data}{flexsurvreg}(x, ...)

\method{axe_env}{flexsurvreg}(x, ...)

\method{axe_fitted}{flexsurvreg}(x, ...)

\method{axe_misc}{flexsurvreg}(x, ...)

\method{predict}{butcher_flexsurvreg}(x, ...)
}
\description{
This is where all the flexsurvreg specific documentation lies.
}
