% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{axe-formula}
\alias{axe-formula}
\alias{axe_env.formula}
\title{Axing formulas.}
\usage{
\method{axe_env}{formula}(x, ...)
}
\arguments{
\item{x}{formula}

\item{...}{any additional arguments related to axing}
}
\value{
axed formula
}
\description{
formulas often wrap environments that carry a lot of unneccesary junk.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(caret)))

data(iris)
train_data <- iris[, 1:4]
train_classes <- iris[, 5]

train_fit <- train(train_data, train_classes,
                   method = "knn",
                   preProcess = c("center", "scale"),
                   tuneLength = 10,
                   trControl = trainControl(method = "cv"))

axe_env(train_fit$modelInfo$prob)

}
