% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{axe-glm}
\alias{axe-glm}
\alias{axe_call.glm}
\alias{axe_data.glm}
\alias{axe_env.glm}
\alias{axe_fitted.glm}
\title{Axing a glm.}
\usage{
\method{axe_call}{glm}(x, verbose = FALSE, ...)

\method{axe_data}{glm}(x, verbose = FALSE, ...)

\method{axe_env}{glm}(x, verbose = FALSE, ...)

\method{axe_fitted}{glm}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed glm object.
}
\description{
glm objects are created from the base \pkg{stats} package.
}
\examples{

cars_glm <- glm(mpg ~ ., data = mtcars)
cleaned_glm <- butcher(cars_glm, verbose = TRUE)

}
