% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{axe-keras}
\alias{axe-keras}
\title{Axing an keras.}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
keras objects are created from the \code{keras} package,
a deep learning framework that can be leveraged to do binary
classification, multiclass classification, regresssion, and
more. This is where all the keras specific documentation lies.
}
