% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{axe-matrix}
\alias{axe-matrix}
\alias{axe_misc.matrix}
\title{Axing matrices.}
\usage{
axe_misc.matrix(x, ...)
}
\arguments{
\item{x}{matrix or matrices}

\item{...}{any additional arguments related to axing}
}
\value{
axed matrix
}
\description{
matrices often store a lot of intermediary training material that is unnecessary
in downstream data analysis functions.
}
