% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{axe-mda}
\alias{axe-mda}
\alias{axe_call.mda}
\alias{axe_env.mda}
\alias{axe_fitted.mda}
\title{Axing a mda.}
\usage{
\method{axe_call}{mda}(x, verbose = FALSE, ...)

\method{axe_env}{mda}(x, verbose = FALSE, ...)

\method{axe_fitted}{mda}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
This is where all the mda specific documentation lies.
}
\examples{
##
## Insert examples to create and axe model object here...
##
}
