% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{axe-model_fit}
\alias{axe-model_fit}
\alias{axe_call.model_fit}
\alias{axe_ctrl.model_fit}
\alias{axe_data.model_fit}
\alias{axe_env.model_fit}
\alias{axe_fitted.model_fit}
\title{Axing an model_fit.}
\usage{
\method{axe_call}{model_fit}(x, verbose = FALSE, ...)

\method{axe_ctrl}{model_fit}(x, verbose = FALSE, ...)

\method{axe_data}{model_fit}(x, verbose = FALSE, ...)

\method{axe_env}{model_fit}(x, verbose = FALSE, ...)

\method{axe_fitted}{model_fit}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
model_fit objects are created from the \code{parsnip} package.
This is where all the model_fit specific documentation lies.
}
