% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnet.R
\name{axe-multnet}
\alias{axe-multnet}
\alias{axe.multnet}
\alias{axe_call.multnet}
\alias{axe_env.multnet}
\alias{axe_fitted.multnet}
\alias{axe_misc.multnet}
\alias{predict.butcher_multnet}
\title{Axing an multnet.}
\usage{
\method{axe}{multnet}(x, ...)

\method{axe_call}{multnet}(x, ...)

\method{axe_env}{multnet}(x, ...)

\method{axe_fitted}{multnet}(x, ...)

\method{axe_misc}{multnet}(x, ...)

\method{predict}{butcher_multnet}(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
axed model object with new butcher class assignment
}
\description{
This is where all the multnet specific documentation lies.

Axe accepts a model object and removes any extraneous
information contained in that object. Generally, this includes the
environments that were carried through from training, the call and
any control parameters that were previously utilized to construct
the model object, as well as the original data and other miscellaneous
parameters that were saved. These parts of the model objects are
removed as they are often not used in any post-processing of the
model object (i.e., prediction).
}
