% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{axe-randomForest}
\alias{axe-randomForest}
\alias{axe_call.randomForest}
\alias{axe_ctrl.randomForest}
\alias{axe_env.randomForest}
\title{Axing an randomForest.}
\usage{
\method{axe_call}{randomForest}(x, verbose = FALSE, ...)

\method{axe_ctrl}{randomForest}(x, verbose = FALSE, ...)

\method{axe_env}{randomForest}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed randomForest object.
}
\description{
randomForest objects are created from the \code{randomForest}
package, which is used to train random forests based on Breiman's
2001 work. The package supports ensembles of classification and
regression trees.
}
