% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{axe-ranger}
\alias{axe-ranger}
\alias{axe_call.ranger}
\alias{axe_fitted.ranger}
\title{Axing an ranger.}
\usage{
\method{axe_call}{ranger}(x, verbose = TRUE, ...)

\method{axe_fitted}{ranger}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
ranger objects are created from the \pkg{ranger} package, which is used as
a means to quickly train random forests. The package supports ensembles
of classification, regression, survival and probability prediction trees.
Given the reliance of post processing functions on the model object, like
\code{importance_pvalues} and \code{treeInfo}, on the first class listed,
the \code{butcher_ranger} class is not appended. This is where all the ranger
specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))

# Load data
set.seed(1234)
split <- initial_split(iris, props = 9/10)
iris_train <- training(split)

# Create model and fit
ranger_fit <- rand_forest(mode = "classification",
                          mtry = 2,
                          trees = 20,
                          min_n = 3) \%>\%
  set_engine("ranger") \%>\%
  fit(Species ~ ., data = iris_train)

butcher(ranger_fit)

}
