% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{axe-recipe}
\alias{axe-recipe}
\alias{axe_env.recipe}
\alias{axe_env.step}
\alias{axe_env.step_arrange}
\alias{axe_env.step_bagimpute}
\alias{axe_env.step_bin2factor}
\alias{axe_env.step_BoxCox}
\alias{axe_env.step_bs}
\alias{axe_env.step_center}
\alias{axe_env.step_classdist}
\alias{axe_env.step_corr}
\alias{axe_env.step_count}
\alias{axe_env.step_date}
\alias{axe_env.step_depth}
\alias{axe_env.step_discretize}
\alias{axe_env.step_downsample}
\alias{axe_env.step_dummy}
\alias{axe_env.step_factor2string}
\alias{axe_env.step_filter}
\alias{axe_env.step_geodist}
\alias{axe_env.step_holiday}
\alias{axe_env.step_hyperbolic}
\alias{axe_env.step_ica}
\alias{axe_env.step_integer}
\alias{axe_env.step_interact}
\alias{axe_env.step_inverse}
\alias{axe_env.step_invlogit}
\alias{axe_env.step_isomap}
\alias{axe_env.step_knnimpute}
\alias{axe_env.step_kpca}
\alias{axe_env.step_lag}
\alias{axe_env.step_lincomb}
\alias{axe_env.step_log}
\alias{axe_env.step_logit}
\alias{axe_env.step_lowerimpute}
\alias{axe_env.step_meanimpute}
\alias{axe_env.step_medianimpute}
\alias{axe_env.step_modeimpute}
\alias{axe_env.step_mutate}
\alias{axe_env.step_naomit}
\alias{axe_env.step_nnmf}
\alias{axe_env.step_novel}
\alias{axe_env.step_num2factor}
\alias{axe_env.step_ns}
\alias{axe_env.step_nzv}
\alias{axe_env.step_ordinalscore}
\alias{axe_env.step_other}
\alias{axe_env.step_pca}
\alias{axe_env.step_pls}
\alias{axe_env.step_poly}
\alias{axe_env.step_range}
\alias{axe_env.step_ratio}
\alias{axe_env.step_regex}
\alias{axe_env.step_relu}
\alias{axe_env.step_rm}
\alias{axe_env.step_rollimpute}
\alias{axe_env.step_shuffle}
\alias{axe_env.step_slice}
\alias{axe_env.step_scale}
\alias{axe_env.step_string2factor}
\alias{axe_env.step_sqrt}
\alias{axe_env.step_spatialsign}
\alias{axe_env.step_unorder}
\alias{axe_env.step_upsample}
\alias{axe_env.step_window}
\alias{axe_env.step_YeoJohnson}
\alias{axe_env.step_zv}
\alias{axe_env.quosure}
\title{Axing a recipe object.}
\usage{
axe_env.recipe(x, ...)

axe_env.step(x, ...)

axe_env.step_arrange(x, ...)

axe_env.step_bagimpute(x, ...)

axe_env.step_bin2factor(x, ...)

axe_env.step_BoxCox(x, ...)

axe_env.step_bs(x, ...)

axe_env.step_center(x, ...)

axe_env.step_classdist(x, ...)

axe_env.step_corr(x, ...)

axe_env.step_count(x, ...)

axe_env.step_date(x, ...)

axe_env.step_depth(x, ...)

axe_env.step_discretize(x, ...)

axe_env.step_downsample(x, ...)

axe_env.step_dummy(x, ...)

axe_env.step_factor2string(x, ...)

axe_env.step_filter(x, ...)

axe_env.step_geodist(x, ...)

axe_env.step_holiday(x, ...)

axe_env.step_hyperbolic(x, ...)

axe_env.step_ica(x, ...)

axe_env.step_integer(x, ...)

axe_env.step_interact(x, ...)

axe_env.step_inverse(x, ...)

axe_env.step_invlogit(x, ...)

axe_env.step_isomap(x, ...)

axe_env.step_knnimpute(x, ...)

axe_env.step_kpca(x, ...)

axe_env.step_lag(x, ...)

axe_env.step_lincomb(x, ...)

axe_env.step_log(x, ...)

axe_env.step_logit(x, ...)

axe_env.step_lowerimpute(x, ...)

axe_env.step_meanimpute(x, ...)

axe_env.step_medianimpute(x, ...)

axe_env.step_modeimpute(x, ...)

axe_env.step_mutate(x, ...)

axe_env.step_naomit(x, ...)

axe_env.step_nnmf(x, ...)

axe_env.step_novel(x, ...)

axe_env.step_num2factor(x, ...)

axe_env.step_ns(x, ...)

axe_env.step_nzv(x, ...)

axe_env.step_ordinalscore(x, ...)

axe_env.step_other(x, ...)

axe_env.step_pca(x, ...)

axe_env.step_pls(x, ...)

axe_env.step_poly(x, ...)

axe_env.step_range(x, ...)

axe_env.step_ratio(x, ...)

axe_env.step_regex(x, ...)

axe_env.step_relu(x, ...)

axe_env.step_rm(x, ...)

axe_env.step_rollimpute(x, ...)

axe_env.step_shuffle(x, ...)

axe_env.step_slice(x, ...)

axe_env.step_scale(x, ...)

axe_env.step_string2factor(x, ...)

axe_env.step_sqrt(x, ...)

axe_env.step_spatialsign(x, ...)

axe_env.step_unorder(x, ...)

axe_env.step_upsample(x, ...)

axe_env.step_window(x, ...)

axe_env.step_YeoJohnson(x, ...)

axe_env.step_zv(x, ...)

axe_env.quosure(x, ...)
}
\arguments{
\item{x}{recipe object}

\item{...}{any additional arguments related to axing}
}
\value{
axed recipe
}
\description{
recipe objects are created from the \code{recipes} package, which is
leveraged for its set of data pre-processing tools. These recipes work
by sequentially defining each pre-processing step. The implementation of
each step, however, results its own class so we bundle all the axe
generics related to recipe objects in general here.
}
