% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{axe-rpart}
\alias{axe-rpart}
\alias{axe.rpart}
\alias{axe_call.rpart}
\alias{axe_ctrl.rpart}
\alias{predict.butcher_rpart}
\title{Axing a rpart.}
\usage{
axe.rpart(x, ...)

\method{axe_call}{rpart}(x, ...)

\method{axe_ctrl}{rpart}(x, ...)

\method{predict}{butcher_rpart}(x, ...)
}
\description{
This is where all the rpart specific documentation lies.
}
