% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsample.R
\name{axe-rsample}
\alias{axe-rsample}
\alias{axe_rsample_data}
\alias{axe_rsample_data.default}
\alias{axe_rsample_data.tune_results}
\alias{axe_rsample_data.rset}
\alias{axe_rsample_data.rsplit}
\alias{axe_rsample_data.three_way_split}
\title{Axing rsample objects}
\usage{
axe_rsample_data(x, verbose = FALSE, ...)

\method{axe_rsample_data}{default}(x, verbose = FALSE, ...)

\method{axe_rsample_data}{tune_results}(x, verbose = FALSE, ...)

\method{axe_rsample_data}{rset}(x, verbose = FALSE, ...)

\method{axe_rsample_data}{rsplit}(x, verbose = FALSE, ...)

\method{axe_rsample_data}{three_way_split}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed rset object.
}
\description{
sclass objects are byproducts of classbagg objects.
}
