% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{axe-spark}
\alias{axe-spark}
\alias{axe_env.ml_model}
\alias{axe_env.ml_pipeline_model}
\alias{axe_env.ml_pipeline_stage}
\alias{axe_misc.ml_model}
\title{Axing a spark object.}
\usage{
axe_env.ml_model(x, ...)

axe_env.ml_pipeline_model(x, ...)

axe_env.ml_pipeline_stage(x, ...)

axe_misc.ml_model(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
spark objects are created from the \code{sparklyr} package,
a \code{R} interface for Apache Spark. This is where all the
spark specific documentation lies.
}
