% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg.R
\name{axe-stanreg}
\alias{axe-stanreg}
\alias{axe_call.stanreg}
\alias{axe_env.stanreg}
\alias{axe_fitted.stanreg}
\title{Axing a stanreg.}
\usage{
\method{axe_call}{stanreg}(x, verbose = TRUE, ...)

\method{axe_env}{stanreg}(x, verbose = TRUE, ...)

\method{axe_fitted}{stanreg}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
This is where all the stanreg specific documentation lies.
}
