% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.penal.R
\name{axe-survreg.penal}
\alias{axe-survreg.penal}
\alias{axe_call.survreg.penal}
\alias{axe_data.survreg.penal}
\alias{axe_env.survreg.penal}
\title{Axing an survreg.penal}
\usage{
\method{axe_call}{survreg.penal}(x, verbose = TRUE, ...)

\method{axe_data}{survreg.penal}(x, verbose = TRUE, ...)

\method{axe_env}{survreg.penal}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
survreg.penal objects are created from the \pkg{survival} package. They
are returned from the \code{survreg} function, representing fitted
parametric survival models. This is where all the survreg.penal specific
documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(survival)))

# Create model and fit
survreg_fit <- surv_reg(mode = "regression", dist = "weibull") \%>\%
  set_engine("survreg") \%>\%
  fit(Surv(time, status) ~ rx + frailty.gaussian(litter, df = 13), data = rats)

butcher(survreg_fit)

}
