% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{axe-survreg}
\alias{axe-survreg}
\alias{axe.survreg}
\alias{axe_call.survreg}
\alias{axe_ctrl.survreg}
\alias{axe_data.survreg}
\alias{axe_env.survreg}
\alias{axe_fitted.survreg}
\alias{axe_misc.survreg}
\alias{predict.butcher_survreg}
\title{Axing an survreg.}
\usage{
\method{axe}{survreg}(x, ...)

\method{axe_call}{survreg}(x, ...)

\method{axe_ctrl}{survreg}(x, ...)

\method{axe_data}{survreg}(x, ...)

\method{axe_env}{survreg}(x, ...)

\method{axe_fitted}{survreg}(x, ...)

\method{axe_misc}{survreg}(x, ...)

\method{predict}{butcher_survreg}(x, ...)
}
\description{
This is where all the survreg specific documentation lies.
}
