% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{axe-terms}
\alias{axe-terms}
\alias{axe_env.terms}
\title{Axing for terms inputs.}
\usage{
\method{axe_env}{terms}(x, ...)
}
\description{
These terms axe functions treat the model object as if
nothing needs to be done.
}
