% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe_call.R
\name{axe_call}
\alias{axe_call}
\title{Axe the call.}
\usage{
axe_call(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without call attribute
}
\description{
Replace the call object attached to modeling objects with a placeholder.
}
\examples{
axe_call(lm_fit)
}
