% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_ctrl}
\alias{axe_ctrl}
\title{Axe controls.}
\usage{
axe_ctrl(x, ...)
}
\arguments{
\item{x}{Model object.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Model object without control tuning parameters from training.
}
\description{
Remove the controls from training attached to modeling objects.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_ctrl")}
}

