% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe_data.R
\name{axe_data}
\alias{axe_data}
\title{Axe data.}
\usage{
axe_data(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without the training data
}
\description{
Remove the training data attached to modeling objects.
}
\examples{
axe_data(kknn_fit)
}
