% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe_env.R
\name{axe_env}
\alias{axe_env}
\title{Axe an environment.}
\usage{
axe_env(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without environment attribut
}
\description{
Remove the environment(s) attached to modeling objects as they are often
not required in the downstream analysis pipeline. Currently, if found,
the environment is completely removed. In future iterations, we might
want to consider replacing the environment object with a NULL.
}
\examples{
axe_env(lm_fit)
}
