% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_misc}
\alias{axe_misc}
\title{Axe miscellaneous.}
\usage{
axe_misc(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
model object without the extra intermediatary units previously stored for training.
}
\description{
Remove the intermediary steps attached from training to modeling objects.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_misc")}
}

