% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{butcher_example}
\alias{butcher_example}
\title{Get path to model object example.}
\usage{
butcher_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If `NULL`, the example files will be listed.}
}
\description{
butcher comes bundled with some example files in its `inst/extdata`
directory. This function was copied from readxl and placed here to
make the instantiated model objects easy to access.
}
\keyword{internal}
