% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find.R
\name{find}
\alias{find}
\title{Find an object.}
\usage{
find(x, item_name = "env")
}
\arguments{
\item{x}{model object}

\item{item_name}{name associated with object component}
}
\value{
location in model object
}
\description{
Locate where a specific component of a object might exist. This
function is restricted in that only items that can be axed will
be found.
}
