% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Locate part of an object.}
\usage{
locate(x, item_name = "env")
}
\arguments{
\item{x}{Model object.}

\item{item_name}{Name associated with object component.}
}
\value{
Location in model object.
}
\description{
Locate where a specific component of a object might exist. This
function is restricted in that only items that can be axed will
be found.
}
