% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Wrapper for predict.}
\usage{
predict(x, ...)
}
\arguments{
\item{x}{axed model object}
}
\value{
output from correct predict function
}
\description{
Since all the axed model objects are assigned to some butcher_ class, in order
for these smaller objects to work downstream with post-training functions
like `predict`, we need to reassign it to its old class.
}
