% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/take_inventory.R
\name{take_inventory}
\alias{take_inventory}
\title{Take inventory.}
\usage{
take_inventory(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
itemized parts of model object
}
\description{
List all the components of a model object to prepare for processing.
}
\examples{
take_inventory(lm_fit)
}
