% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\alias{memory_released}
\alias{assess_object}
\title{Console Messages}
\usage{
memory_released(og, butchered)

assess_object(og, butchered, disabled = NULL, class_added = TRUE)
}
\arguments{
\item{og}{Original model object.}

\item{butchered}{Butchered model object.}

\item{disabled}{String of disabled functions as a result of axing.}

\item{class_added}{Boolean that defaults to true when butcher class is added.}
}
\description{
These functions leverage the \code{ui.R} as provided in the
\pkg{usethis} package. Original reference here:
\url{https://github.com/r-lib/usethis/blob/master/R/ui.R}.
These console messages are created such that the user is
aware of the effects of removing specific components from
the model object.
}
