% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weigh.R
\name{weigh}
\alias{weigh}
\title{Weigh the object.}
\usage{
weigh(x, threshold = 2, units = "MB")
}
\arguments{
\item{x}{model object}

\item{threshold}{cutoff memory level}

\item{units}{defaults to MB}
}
\value{
tibble
}
\description{
Evaluate the size of each element contained in a model object.
}
\examples{
simulate_x <- matrix(runif(1e+6), ncol = 2)
simulate_y <- runif(dim(simulate_x)[1])
lm_out <- lm(simulate_y ~ simulate_x)
weigh(lm_out)
}
