% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c5.R
\name{axe-C5.0}
\alias{axe-C5.0}
\alias{axe_call.C5.0}
\alias{axe_ctrl.C5.0}
\alias{axe_fitted.C5.0}
\title{Axing a C5.0.}
\usage{
\method{axe_call}{C5.0}(x, ...)

\method{axe_ctrl}{C5.0}(x, ...)

\method{axe_fitted}{C5.0}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
C5.0 objects are created from the \code{C50} package, which provides an
interface to the C5.0 classification model. The models that can be
generated include basic tree-based models as well as rule-based models.
This is where all the C5.0 specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(tidymodels)))
suppressWarnings(suppressMessages(library(rpart)))

# Load data
set.seed(1234)
split <- initial_split(kyphosis, props = 9/10)
spine_train <- training(split)

# Create model and fit
c5_fit <- decision_tree(mode = "classification") \%>\%
  set_engine("C5.0") \%>\%
  fit(Kyphosis ~ ., data = spine_train)

butcher(c5_fit)

}
