% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classbagg.R
\name{axe-classbagg}
\alias{axe-classbagg}
\alias{axe_call.classbagg}
\alias{axe_data.classbagg}
\alias{axe_env.classbagg}
\title{Axing a classbagg object.}
\usage{
\method{axe_call}{classbagg}(x, verbose = TRUE, ...)

\method{axe_data}{classbagg}(x, verbose = TRUE, ...)

\method{axe_env}{classbagg}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Classbagg object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed classbagg object.
}
\description{
This is where all the classbagg specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(ipred)))
suppressWarnings(suppressMessages(library(rpart)))
suppressWarnings(suppressMessages(library(MASS)))

# Load data
data("GlaucomaM", package = "TH.data")

classbagg_fit <- bagging(Class ~ ., data = GlaucomaM, coob = TRUE)

butcher(classbagg_fit)
}
