% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earth.R
\name{axe-earth}
\alias{axe-earth}
\alias{axe_call.earth}
\alias{axe_data.earth}
\alias{axe_fitted.earth}
\title{Axing an earth object.}
\usage{
\method{axe_call}{earth}(x, verbose = TRUE, ...)

\method{axe_data}{earth}(x, verbose = TRUE, ...)

\method{axe_fitted}{earth}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
earth objects are created from the \pkg{earth} package, which
is leveraged to do multivariate adapative regression splines.
This is where all the earth specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))

# Create model and fit
earth_fit <- mars(mode = "regression") \%>\%
  set_engine("earth") \%>\%
  fit(Volume ~ ., data = trees)

butcher(earth_fit)

}
