% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet.R
\name{axe-elnet}
\alias{axe-elnet}
\alias{axe.elnet}
\alias{axe_call.elnet}
\alias{axe_env.elnet}
\alias{axe_fitted.elnet}
\alias{predict.butcher_elnet}
\title{Axing an elnet.}
\usage{
axe.elnet(x, ...)

\method{axe_call}{elnet}(x, ...)

\method{axe_env}{elnet}(x, ...)

\method{axe_fitted}{elnet}(x, ...)

\method{predict}{butcher_elnet}(x, ...)
}
\description{
This is where all the elnet specific documentation lies.
}
