% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elnet.R
\name{axe-elnet}
\alias{axe-elnet}
\alias{axe_call.elnet}
\title{Axing an elnet.}
\usage{
\method{axe_call}{elnet}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
This is where all the elnet specific documentation lies.
}
\details{
Note: the elnet object is one of the few model objects in which
there is no environment to axe.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))

# Load data
split <- initial_split(mtcars, props = 9/10)
car_train <- training(split)
car_test  <- testing(split)

# Create model and fit
elnet_fit <- linear_reg(mixture = 0, penalty = 0.1) \%>\%
  set_engine("glmnet") \%>\%
  fit_xy(x = car_train[, 2:11], y = car_train[, 1, drop = FALSE])

butcher(elnet_fit)

}
