% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{axe-formula}
\alias{axe-formula}
\alias{axe_env.formula}
\title{Axing formulas.}
\usage{
\method{axe_env}{formula}(x, ...)
}
\arguments{
\item{x}{formula}

\item{...}{any additional arguments related to axing}
}
\value{
axed formula
}
\description{
formulas often wrap environments that carry a lot of unneccesary junk.
}
