% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras.R
\name{axe-keras}
\alias{axe-keras}
\alias{axe.keras}
\alias{axe_call.keras}
\alias{axe_env.keras}
\alias{axe_fitted.keras}
\alias{predict.butcher_keras}
\title{Axing an keras.}
\usage{
axe.keras(x, ...)

\method{axe_call}{keras}(x, ...)

\method{axe_env}{keras}(x, ...)

\method{axe_fitted}{keras}(x, ...)

\method{predict}{butcher_keras}(x, ...)
}
\description{
This is where all the keras specific documentation lies.
}
