% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kknn.R
\name{axe-kknn}
\alias{axe-kknn}
\alias{axe.kknn}
\alias{axe_call.kknn}
\alias{axe_env.kknn}
\alias{axe_fitted.kknn}
\alias{axe_misc.kknn}
\alias{predict.butcher_kknn}
\title{Axing an kknn.}
\usage{
axe.kknn(x, ...)

\method{axe_call}{kknn}(x, ...)

\method{axe_env}{kknn}(x, ...)

\method{axe_fitted}{kknn}(x, ...)

\method{axe_misc}{kknn}(x, ...)

\method{predict}{butcher_kknn}(x, ...)
}
\description{
This is where all the kknn specific documentation lies.
}
