% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ksvm.R
\name{axe-ksvm}
\alias{axe-ksvm}
\title{Axing a ksvm object.}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
This is where all the ksvm specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(kernlab)))

# Load data
data(spam)

# Create model and fit
ksvm_class <- svm_poly(mode = "classification") \%>\%
  set_engine("kernlab", kernel = "rbfdot") \%>\%
  fit(type ~ ., data = spam)

butcher(ksvm_class)

}
