% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.R
\name{axe-lm}
\alias{axe-lm}
\alias{axe_call.lm}
\alias{axe_env.lm}
\alias{axe_fitted.lm}
\title{Axing an lm.}
\usage{
\method{axe_call}{lm}(x, verbose = TRUE, ...)

\method{axe_env}{lm}(x, verbose = TRUE, ...)

\method{axe_fitted}{lm}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
This is where all the lm specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))

# Load data
split <- initial_split(mtcars, props = 9/10)
car_train <- training(split)
car_test  <- testing(split)

# Create model and fit
lm_fit <- linear_reg() \%>\%
  set_engine("lm") \%>\%
  fit(mpg ~ ., data = car_train)

butcher(lm_fit)

}
