% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmnet.R, R/lm.R
\name{axe-lm}
\alias{axe-lm}
\alias{axe.lm}
\alias{axe_call.lm}
\alias{axe_env.lm}
\alias{axe_env.fitted}
\alias{predict.butcher_lm}
\alias{axe_fitted.lm}
\title{Axing an lm.}
\usage{
axe.lm(x, ...)

\method{axe_call}{lm}(x, ...)

\method{axe_env}{lm}(x, ...)

\method{axe_env}{fitted}(x, ...)

\method{predict}{butcher_lm}(x, ...)

\method{axe_call}{lm}(x, ...)

\method{axe_env}{lm}(x, ...)

\method{axe_fitted}{lm}(x, ...)

\method{predict}{butcher_lm}(x, ...)
}
\description{
This is where all the lm specific documentation lies.

This is where all the lm specific documentation lies.
}
\examples{
# Make a lm that has a lot of extra stuff in its environment
make_an_lm <- function() {
  x <- rep(1L, times = 10e6)
  lm(1 ~ 1)
}
lm_object <- make_an_lm()
lm_axed <- axe(lm)
}
