% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mda.R
\name{axe-mda}
\alias{axe-mda}
\alias{axe_call.mda}
\alias{axe_env.mda}
\alias{axe_fitted.mda}
\title{Axing a mda.}
\usage{
\method{axe_call}{mda}(x, verbose = FALSE, ...)

\method{axe_env}{mda}(x, verbose = FALSE, ...)

\method{axe_fitted}{mda}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed mda object.
}
\description{
mda objects are created from the \pkg{mda} package, leveraged to
carry out mixture discriminant analysis.
}
\examples{
\dontshow{if (rlang::is_installed("mda")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(mda)

fit <- mda(Species ~ ., data = iris)
out <- butcher(fit, verbose = TRUE)

# Another mda object
data(glass)
wrapped_mda <- function() {
  some_junk_in_environment <- runif(1e6)
  fit <- mda(Type ~ ., data = glass)
  return(fit)
}

lobstr::obj_size(wrapped_mda())
lobstr::obj_size(butcher(wrapped_mda()))
\dontshow{\}) # examplesIf}
}
