% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{axe-model_fit}
\alias{axe-model_fit}
\alias{axe_call.model_fit}
\alias{axe_ctrl.model_fit}
\alias{axe_data.model_fit}
\alias{axe_env.model_fit}
\alias{axe_fitted.model_fit}
\alias{axe_misc.model_fit}
\title{Axing an model_fit.}
\usage{
axe_call.model_fit(x, ...)

axe_ctrl.model_fit(x, ...)

axe_data.model_fit(x, ...)

axe_env.model_fit(x, ...)

axe_fitted.model_fit(x, ...)

axe_misc.model_fit(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
model_fit objects are created from the \code{parsnip} package.
This is where all the model_fit specific documentation lies.
}
