% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R
\name{axe-model_fit}
\alias{axe-model_fit}
\alias{axe.model_fit}
\alias{axe_call.model_fit}
\alias{axe_ctrl.model_fit}
\alias{axe_data.model_fit}
\alias{axe_env.model_fit}
\alias{axe_fitted.model_fit}
\alias{axe_misc.model_fit}
\alias{predict.butcher_model_fit}
\title{Axing an model_fit.}
\usage{
axe.model_fit(x, ...)

\method{axe_call}{model_fit}(x, ...)

\method{axe_ctrl}{model_fit}(x, ...)

\method{axe_data}{model_fit}(x, ...)

\method{axe_env}{model_fit}(x, ...)

\method{axe_fitted}{model_fit}(x, ...)

\method{axe_misc}{model_fit}(x, ...)

\method{predict}{butcher_model_fit}(x, ...)
}
\description{
This is where all the model_fit specific documentation lies.
}
