% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnet.R
\name{axe-multnet}
\alias{axe-multnet}
\alias{axe_call.multnet}
\title{Axing an multnet.}
\usage{
\method{axe_call}{multnet}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
This is where all the multnet specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))

# Load data
set.seed(1234)
predictrs <- matrix(rnorm(100*20), ncol = 20)
response <- as.factor(sample(1:4, 100, replace = TRUE))

# Create model and fi
multnet_fit <- multinom_reg() \%>\%
  set_engine("glmnet") \%>\%
  fit_xy(x = predictrs, y = response)

butcher(multnet_fit)

}
