% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multnet.R
\name{axe-multnet}
\alias{axe-multnet}
\alias{axe_call.multnet}
\alias{axe_misc.multnet}
\alias{predict.butcher_multnet}
\title{Axing an multnet.}
\usage{
\method{axe_call}{multnet}(x, ...)

\method{axe_misc}{multnet}(x, ...)

\method{predict}{butcher_multnet}(x, ...)
}
\description{
This is where all the multnet specific documentation lies.
}
\details{
Note: Since multnet object is derived from the \code{glmnet} package it is
one of the few model objects in which there is no environment to axe.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(tidymodels)))

# Load data
set.seed(1234)
predictrs <- matrix(rnorm(100*20), ncol = 20)
response <- as.factor(sample(1:4, 100, replace = TRUE))

# Create model and fi
multnet_fit <- multinom_reg() \%>\%
  set_engine("glmnet") \%>\%
  fit_xy(x = predictrs, y = response)

# Axe
axe(multnet_fit)
}
