% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnet.R
\name{axe-nnet}
\alias{axe-nnet}
\alias{axe_call.nnet}
\alias{axe_env.nnet}
\alias{axe_fitted.nnet}
\title{Axing a nnet.}
\usage{
\method{axe_call}{nnet}(x, ...)

\method{axe_env}{nnet}(x, ...)

\method{axe_fitted}{nnet}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
nnet objects are created from the \pkg{nnet} package.
This is where all the nnet specific documentation lies.
}
