% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomForest.R
\name{axe-randomForest}
\alias{axe-randomForest}
\alias{axe.randomForest}
\alias{axe_call.randomForest}
\alias{axe_ctrl.randomFoest}
\alias{axe_env.randomForest}
\alias{axe_fitted.randomForest}
\alias{axe_misc.randomForest}
\alias{predict.butcher_randomForest}
\title{Axing an randomForest.}
\usage{
\method{axe}{randomForest}(x, ...)

\method{axe_call}{randomForest}(x, ...)

\method{axe_ctrl}{randomFoest}(x, ...)

\method{axe_env}{randomForest}(x, ...)

\method{axe_fitted}{randomForest}(x, ...)

\method{axe_misc}{randomForest}(x, ...)

\method{predict}{butcher_randomForest}(x, ...)
}
\description{
This is where all the randomForest specific documentation lies.
}
