% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{axe-ranger}
\alias{axe-ranger}
\alias{axe.ranger}
\alias{axe_call.ranger}
\alias{axe_ctrl.ranger}
\alias{axe_env.ranger}
\alias{axe_fitted.ranger}
\alias{axe_misc.ranger}
\alias{predict.butcher_ranger}
\title{Axing an ranger.}
\usage{
\method{axe}{ranger}(x, ...)

\method{axe_call}{ranger}(x, ...)

\method{axe_ctrl}{ranger}(x, ...)

\method{axe_env}{ranger}(x, ...)

\method{axe_fitted}{ranger}(x, ...)

\method{axe_misc}{ranger}(x, ...)

\method{predict}{butcher_ranger}(x, ...)
}
\description{
This is where all the ranger specific documentation lies.
}
