% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger.R
\name{axe-ranger}
\alias{axe-ranger}
\alias{axe_call.ranger}
\alias{axe_fitted.ranger}
\title{Axing an ranger.}
\usage{
\method{axe_call}{ranger}(x, ...)

\method{axe_fitted}{ranger}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
ranger objects are created from the \code{ranger} package, which is used as
a means to quickly train random forests. The package supports ensembles
of classification, regression, survival and probability prediction trees.
This is where all the ranger specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(tidymodels)))

# Load data
set.seed(1234)
split <- initial_split(iris, props = 9/10)
iris_train <- training(split)

# Create model and fit
ranger_fit <- rand_forest(mode = "classification",
                          mtry = 2,
                          trees = 20,
                          min_n = 3) \%>\%
  set_engine("ranger") \%>\%
  fit(Species ~ ., data = iris_train)

butcher(ranger_fit)

}
