% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipe.R
\name{axe-recipe}
\alias{axe-recipe}
\alias{axe_env.recipe}
\title{Axing a recipe object.}
\usage{
\method{axe_env}{recipe}(x, ...)
}
\arguments{
\item{x}{recipe object}

\item{...}{any additional arguments related to axing}
}
\value{
axed recipe
}
\description{
This is where all the recipe specific documentation lies.
}
