% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart.R
\name{axe-rpart}
\alias{axe-rpart}
\alias{axe_call.rpart}
\alias{axe_ctrl.rpart}
\alias{axe_data.rpart}
\alias{axe_env.rpart}
\title{Axing a rpart.}
\usage{
\method{axe_call}{rpart}(x, verbose = TRUE, ...)

\method{axe_ctrl}{rpart}(x, verbose = TRUE, ...)

\method{axe_data}{rpart}(x, verbose = TRUE, ...)

\method{axe_env}{rpart}(x, verbose = TRUE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{TRUE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
rpart objects are created from the \pkg{rpart} package, which
is used for recursive partitioning for classification, regression and
survival trees. This is where all the rpart specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(rsample)))
suppressWarnings(suppressMessages(library(rpart)))

# Load data
set.seed(1234)
split <- initial_split(mtcars, props = 9/10)
car_train <- training(split)

# Create model and fit
rpart_fit <- decision_tree(mode = "regression") \%>\%
  set_engine("rpart") \%>\%
  fit(mpg ~ ., data = car_train, minsplit = 5, cp = 0.1)

butcher(rpart_fit)

}
