% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe-rsample}
\alias{axe-rsample}
\alias{axe_rsample_data}
\alias{axe_rsample_indicators}
\alias{axe_rsample_indicators.default}
\title{Axe rsample objects.}
\usage{
axe_rsample_data(x, verbose = FALSE, ...)

axe_rsample_indicators(x, verbose = FALSE, ...)

\method{axe_rsample_indicators}{default}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
An update object with no data or integer indicators.
}
\description{
Replace the splitting and resampling objects with a placeholder.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_call")}
}

