% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclass.R
\name{axe-sclass}
\alias{axe-sclass}
\alias{axe_call.sclass}
\alias{axe_env.sclass}
\title{Axing a sclass object.}
\usage{
axe_call.sclass(x, ...)

axe_env.sclass(x, ...)
}
\arguments{
\item{x}{sclass object}

\item{...}{any additional arguments related to axing}
}
\value{
axed sclass object
}
\description{
sclass objects are byproducts of classbagg objects. Since there are
axe functions specific to this class, we keep all documentation
related to sclass objects here.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(ipred)))
suppressWarnings(suppressMessages(library(rpart)))
suppressWarnings(suppressMessages(library(MASS)))

# Load data
data("GlaucomaM", package = "TH.data")

classbagg_fit <- bagging(Class ~ ., data = GlaucomaM, coob = TRUE)

butcher(classbagg_fit$mtrees[[1]])
}
