% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sclass.R
\name{axe-sclass}
\alias{axe-sclass}
\alias{axe_call.sclass}
\alias{axe_env.sclass}
\title{Axing a sclass object.}
\usage{
\method{axe_call}{sclass}(x, verbose = FALSE, ...)

\method{axe_env}{sclass}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{sclass object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed sclass object.
}
\description{
sclass objects are byproducts of classbagg objects. Since there are
axe functions specific to this class, there is where all documentation
related to sclass objects lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(ipred)))
suppressWarnings(suppressMessages(library(rpart)))
suppressWarnings(suppressMessages(library(MASS)))

# Load data
data("GlaucomaM", package = "TH.data")

classbagg_fit <- bagging(Class ~ ., data = GlaucomaM, coob = TRUE)

butcher(classbagg_fit$mtrees[[1]])
}
