% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spark.R
\name{axe-spark}
\alias{axe-spark}
\alias{axe_env.ml_model}
\alias{axe_env.ml_pipeline_model}
\alias{axe_env.ml_pipeline_stage}
\title{Axing a spark object.}
\usage{
\method{axe_env}{ml_model}(x, ...)

\method{axe_env}{ml_pipeline_model}(x, ...)

\method{axe_env}{ml_pipeline_stage}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
spark objects are created from the \code{sparklyr} package,
a \code{R} interface for Apache Spark. This is where all the
spark specific documentation lies.
}
