% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg.R
\name{axe-stanreg}
\alias{axe-stanreg}
\alias{axe_call.stanreg}
\alias{axe_env.stanreg}
\alias{axe_fitted.stanreg}
\title{Axing a stanreg.}
\usage{
\method{axe_call}{stanreg}(x, verbose = FALSE, ...)

\method{axe_env}{stanreg}(x, verbose = FALSE, ...)

\method{axe_fitted}{stanreg}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed stanreg object.
}
\description{
stanreg objects are created from the \pkg{rstanarm} package, leveraged
to do Bayesian regression modeling with \pkg{stan}.
}
\examples{
\dontrun{
library(rstanarm)
fit <- stan_glm(mpg ~ wt, data = mtcars, algorithm = "optimizing")
butcher(fit)
}
}
