% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg.R
\name{axe-stanreg}
\alias{axe-stanreg}
\alias{axe.stanreg}
\alias{axe_call.stanreg}
\alias{axe_fitted.stanreg}
\alias{predict.butcher_stanreg}
\title{Axing a stanreg.}
\usage{
axe.stanreg(x, ...)

\method{axe_call}{stanreg}(x, ...)

\method{axe_fitted}{stanreg}(x, ...)

\method{predict}{butcher_stanreg}(x, ...)
}
\description{
This is where all the stanreg specific documentation lies.
}
