% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanreg.R
\name{axe-stanreg}
\alias{axe-stanreg}
\alias{axe_call.stanreg}
\alias{axe_env.stanreg}
\alias{axe_fitted.stanreg}
\title{Axing a stanreg.}
\usage{
\method{axe_call}{stanreg}(x, ...)

\method{axe_env}{stanreg}(x, ...)

\method{axe_fitted}{stanreg}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
This is where all the stanreg specific documentation lies.
}
