% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{axe-survreg}
\alias{axe-survreg}
\alias{axe_call.survreg}
\alias{axe_data.survreg}
\alias{axe_env.survreg}
\title{Axing an survreg.}
\usage{
\method{axe_call}{survreg}(x, verbose = FALSE, ...)

\method{axe_data}{survreg}(x, verbose = FALSE, ...)

\method{axe_env}{survreg}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
survreg objects are created from the \pkg{survival} package. They
are returned from the \code{survreg} function, representing fitted
parametric survival models. This is where all the survreg specific
documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(flexsurv)))

# Create model and fit
survreg_fit <- surv_reg(mode = "regression", dist = "weibull") \%>\%
  set_engine("survreg") \%>\%
  fit(Surv(futime, fustat) ~ 1, data = ovarian)

butcher(survreg_fit)

}
