% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survreg.R
\name{axe-survreg}
\alias{axe-survreg}
\alias{axe_call.survreg}
\alias{axe_data.survreg}
\alias{axe_env.survreg}
\title{Axing an survreg.}
\usage{
\method{axe_call}{survreg}(x, ...)

\method{axe_data}{survreg}(x, ...)

\method{axe_env}{survreg}(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object
}
\description{
survreg objects are created from the \code{survival} package. They
are returned from the \code{survreg} function, representing fitted
parametric survival models. This is where all the survreg specific
documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(parsnip)))
suppressWarnings(suppressMessages(library(tidymodels)))
suppressWarnings(suppressMessages(library(flexsurv)))

# Create model and fit
survreg_fit <- surv_reg(mode = "regression", dist = "weibull") \%>\%
  set_engine("survreg") \%>\%
  fit(Surv(futime, fustat) ~ 1, data = ovarian)

butcher(survreg_fit)

}
