% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terms.R
\name{axe-terms}
\alias{axe-terms}
\alias{axe_env.terms}
\title{Axing for terms inputs.}
\usage{
\method{axe_env}{terms}(x, ...)
}
\arguments{
\item{x}{Terms object.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed terms object.
}
\description{
Generics related to axing objects of the term class.
}
