% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train.R
\name{axe-train}
\alias{axe-train}
\alias{axe_call.train}
\alias{axe_ctrl.train}
\alias{axe_data.train}
\alias{axe_env.train}
\alias{axe_fitted.train}
\title{Axing a train object.}
\usage{
axe_call.train(x, ...)

axe_ctrl.train(x, ...)

axe_data.train(x, ...)

axe_env.train(x, ...)

axe_fitted.train(x, ...)
}
\arguments{
\item{x}{train object}

\item{...}{any additional arguments related to axing}
}
\value{
axed train object
}
\description{
This is where all the train specific documentation lies.
}
\examples{
# Load libraries
suppressWarnings(suppressMessages(library(caret)))

data(iris)
train_data <- iris[, 1:4]
train_classes <- iris[, 5]

train_fit <- train(train_data, train_classes,
                   method = "knn",
                   preProcess = c("center", "scale"),
                   tuneLength = 10,
                   trControl = trainControl(method = "cv"))

out <- butcher(train_fit)

}
