% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{axe-xgb.Booster}
\alias{axe-xgb.Booster}
\alias{axe_call.xgb.Booster}
\alias{axe_ctrl.xgb.Booster}
\alias{axe_env.xgb.Booster}
\alias{axe_fitted.xgb.Booster}
\title{Axing a xgb.Booster.}
\usage{
\method{axe_call}{xgb.Booster}(x, verbose = FALSE, ...)

\method{axe_ctrl}{xgb.Booster}(x, verbose = FALSE, ...)

\method{axe_env}{xgb.Booster}(x, verbose = FALSE, ...)

\method{axe_fitted}{xgb.Booster}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{Model object.}

\item{verbose}{Print information each time an axe method is executed
that notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object.
}
\description{
xgb.Booster objects are created from the \pkg{xgboost} package,
which provides efficient and scalable implementations of gradient
boosted decision trees. Given the reliance of post processing
functions on the model object, like \code{xgb.Booster.complete},
on the first class listed, the \code{butcher_xgb.Booster} class is
not appended. This is where all the xgb.Booster specific documentation
lies.
}
\examples{
library(xgboost)
data(agaricus.train)
data(agaricus.test)
bst <- xgboost(data = agaricus.train$data,
               label = agaricus.train$label,
               max.depth = 2,
               eta = 1,
               nthread = 2,
               nrounds = 2,
               objective = "binary:logistic")

butcher(bst)
}
