% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe}
\alias{axe}
\title{Axe an object.}
\usage{
axe(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
axed model object with new butcher subclass assignment
}
\description{
Reduce the size of a model object so that it takes up less memory on disk.
Currently the model object is stripped down to the point that only the
minimal components necessary for the `predict` function to work remain.
Future adjustments to this function will be needed to avoid removal of
model fit components to ensure it works with other downstream functions.
}
