% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_call}
\alias{axe_call}
\title{Axe a call.}
\usage{
axe_call(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without call attribute
}
\description{
Replace the call object attached to modeling objects with a placeholder.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_call")}
}

