% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_ctrl}
\alias{axe_ctrl}
\title{Axe controls.}
\usage{
axe_ctrl(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without control tuning parameters for training.
}
\description{
Remove the controls attached to modeling objects.
}
\examples{
axe_ctrl(treereg_fit)
}
