% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_data}
\alias{axe_data}
\title{Axe data.}
\usage{
axe_data(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Model object without the training data
}
\description{
Remove the training data attached to modeling objects.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_data")}
}

