% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_env}
\alias{axe_env}
\title{Axe an environment.}
\usage{
axe_env(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{other arguments passed to axe methods}
}
\value{
model object with empty environments
}
\description{
Remove the environment(s) attached to modeling objects as they are often
not required in the downstream analysis pipeline. Currently, if found,
the environment is replaced with rlang::empty_env.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_env")}
}

