% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_env}
\alias{axe_env}
\title{Axe an environment.}
\usage{
axe_env(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Model object with empty environments.
}
\description{
Remove the environment(s) attached to modeling objects as they are
not required in the downstream analysis pipeline. If found,
the environment is replaced with \code{rlang::base_env()}.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_env")}
}

