% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_fitted}
\alias{axe_fitted}
\title{Axe fitted values.}
\usage{
axe_fitted(x, ...)
}
\arguments{
\item{x}{model object}

\item{...}{any additional arguments related to axing}
}
\value{
model object without the fitted values
}
\description{
Remove the fitted values attached to modeling objects.
}
\section{Methods}{

\Sexpr[stage=render,results=rd]{butcher:::methods_rd("axe_fitted")}
}

