% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_fitted}
\alias{axe_fitted}
\title{Axe fitted.}
\usage{
axe_fitted(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without the fitted values
}
\description{
Remove the fitted values attached to modeling objects.
}
\examples{
axe_fitted(kknn_fit)
}
