% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{axe_misc}
\alias{axe_misc}
\title{Axe miscellaneous.}
\usage{
axe_misc(x, ...)
}
\arguments{
\item{x}{model object}
}
\value{
model object without the extra intermediatary units previously stored for training.
}
\description{
Remove the intermediary units attached to modeling objects.
}
\examples{
axe_misc(glmnet_multi)
}
