% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/axe-generics.R
\name{butcher}
\alias{butcher}
\title{Butcher an object.}
\usage{
butcher(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{A model object.}

\item{verbose}{Print information each time an axe method is executed.
Notes how much memory is released and what functions are
disabled. Default is \code{FALSE}.}

\item{...}{Any additional arguments related to axing.}
}
\value{
Axed model object with new butcher subclass assignment.
}
\description{
Reduce the size of a model object so that it takes up less memory on
disk. Currently, the model object is stripped down to the point that
only the minimal components necessary for the \code{predict} function
to work remain. Future adjustments to this function will be needed to
avoid removal of model fit components to ensure it works with other
downstream functions.
}
