% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new_model_butcher}
\alias{new_model_butcher}
\title{New axe functions for a modeling object.}
\usage{
new_model_butcher(model_class, package_name, open = interactive())
}
\arguments{
\item{model_class}{A string that captures the class name of
the new model object.}

\item{package_name}{A string that captures the package name
from which the new model is made.}

\item{open}{Check if user is in interactive mode, and if so,
opens the new files for editing.}
}
\description{
\code{new_model_butcher()} will instantiate the following to help
us develop new axe functions around removing parts of a new
modeling object:
\itemize{
\item Add modeling package to \code{Suggests}
\item Generate and populate an axe file under \code{R/}
\item Generate and populate an test file under \code{testthat/}
}
}
