% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\alias{memory_released}
\alias{assess_object}
\title{Console Messages}
\usage{
memory_released(og, butchered)

assess_object(og, butchered)
}
\arguments{
\item{og}{Original model object.}

\item{butchered}{Butchered model object.}
}
\description{
These console messages are created such that the user is
aware of the effects of removing specific components from
the model object.
}
\keyword{internal}
